상수 대문자처리, 입출력요소 제거
